Save the ability table to Connection.
