--- core/src/main/java/com/alibaba/nacos/core/ability/inte/ClusterAbilityControlSupport.java
@@ -1,46 +1,0 @@
-/*
- * Copyright 1999-2022 Alibaba Group Holding Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.alibaba.nacos.core.ability.inte;
-
-import com.alibaba.nacos.api.ability.constant.AbilityKey;
-import com.alibaba.nacos.api.ability.constant.AbilityStatus;
-import com.alibaba.nacos.common.ability.inter.AbilityControlManager;
-
-import java.util.Map;
-
-/**.
- * @author Daydreamer
- * @description It provides the capability to manage the AbilityTable in cluster for the {@link AbilityControlManager}
- * @date 2022/8/10 23:18
- **/
-public interface ClusterAbilityControlSupport {
-    
-    /**.
-     * Return the cluster abilities.
-     *
-     * @return the cluster abilities.
-     */
-    Map<AbilityKey, Boolean> getClusterAbility();
-    
-    /**.
-     * Whether all the servers currently connected support a certain capability
-     *
-     * @param abilityKey ability key
-     * @return whether it is turn on
-     */
-    AbilityStatus isClusterEnableAbilityNow(AbilityKey abilityKey);
-}
