+++ core/src/test/java/com/alibaba/nacos/core/ability/AbilityControlManagerTest.java
@@ -17,8 +17,6 @@
 package com.alibaba.nacos.core.ability;
 
 import com.alibaba.nacos.api.ability.constant.AbilityKey;
-import com.alibaba.nacos.api.ability.constant.AbilityStatus;
-import com.alibaba.nacos.api.ability.entity.AbilityTable;
 import com.alibaba.nacos.common.ability.handler.HandlerMapping;
 import org.junit.Assert;
 import org.junit.Before;
@@ -40,99 +38,9 @@ public class AbilityControlManagerTest {
     public void inject() {
         Map<AbilityKey, Boolean> newTable = new HashMap<>();
         newTable.put(AbilityKey.TEST_1, true);
-
-        Map<AbilityKey, Boolean> cluster = new HashMap<>();
-        cluster.put(AbilityKey.TEST_1, true);
-        serverAbilityControlManager.setClusterAbility(cluster);
         serverAbilityControlManager.setCurrentSupportingAbility(newTable);
     }
     
-    @Test
-    public void testClientAdd() {
-        Map<AbilityKey, Boolean> newTable = new HashMap<>();
-        newTable.put(AbilityKey.TEST_2, true);
-        newTable.put(AbilityKey.TEST_1, true);
-        AbilityTable table = new AbilityTable();
-        table.setConnectionId("test-00001");
-        table.setAbility(newTable);
-        table.setServer(true);
-        serverAbilityControlManager.addNewTable(table);
-        Assert.assertEquals(AbilityStatus.NOT_SUPPORTED, serverAbilityControlManager.isSupport("test-00001", AbilityKey.TEST_2));
-        Assert.assertEquals(AbilityStatus.SUPPORTED, serverAbilityControlManager.isSupport("test-00001", AbilityKey.TEST_1));
-    }
-    
-    @Test
-    public void testServerAdd() {
-        Map<AbilityKey, Boolean> newTable = new HashMap<>();
-        newTable.put(AbilityKey.TEST_2, true);
-        newTable.put(AbilityKey.TEST_1, true);
-        AbilityTable table = new AbilityTable();
-        table.setConnectionId("test-00001");
-        table.setAbility(newTable);
-        table.setServer(true);
-        serverAbilityControlManager.addNewTable(table);
-        Assert.assertEquals(AbilityStatus.NOT_SUPPORTED, serverAbilityControlManager.isSupport("test-00001", AbilityKey.TEST_2));
-        Assert.assertEquals(AbilityStatus.SUPPORTED, serverAbilityControlManager.isSupport("test-00001", AbilityKey.TEST_1));
-
-        Map<AbilityKey, Boolean> otherServer = new HashMap<>();
-        otherServer.put(AbilityKey.TEST_2, true);
-        otherServer.put(AbilityKey.TEST_1, false);
-        AbilityTable otherServerTable = new AbilityTable();
-        otherServerTable.setConnectionId("test-00000");
-        otherServerTable.setAbility(otherServer);
-        otherServerTable.setServer(true);
-        serverAbilityControlManager.addNewTable(otherServerTable);
-
-        Map<AbilityKey, Boolean> clientTa = new HashMap<>();
-        clientTa.put(AbilityKey.TEST_2, true);
-        clientTa.put(AbilityKey.TEST_1, false);
-        AbilityTable clientTable = new AbilityTable();
-        clientTable.setConnectionId("test-00002");
-        clientTable.setAbility(clientTa);
-        clientTable.setServer(false);
-        serverAbilityControlManager.addNewTable(clientTable);
-        
-        // if not support
-        AbilityTable serverTable = new AbilityTable();
-        serverTable.setConnectionId("test-001231");
-        serverTable.setServer(true);
-        serverAbilityControlManager.addNewTable(serverTable);
-        // unknown because not support
-        Assert.assertEquals(serverAbilityControlManager.isClusterEnableAbilityNow(AbilityKey.TEST_1), AbilityStatus.UNKNOWN);
-        Assert.assertEquals(serverAbilityControlManager.getServerNotSupportAbility().size(), 1);
-        Assert.assertTrue(serverAbilityControlManager.getServerNotSupportAbility().contains("test-001231"));
-        
-        AbilityTable serverTable1 = new AbilityTable();
-        serverTable1.setConnectionId("test-001231231");
-        serverTable1.setServer(true);
-        serverAbilityControlManager.addNewTable(serverTable1);
-        // unknown because not support
-        Assert.assertEquals(serverAbilityControlManager.isClusterEnableAbilityNow(AbilityKey.TEST_1), AbilityStatus.UNKNOWN);
-        Assert.assertEquals(serverAbilityControlManager.getServerNotSupportAbility().size(), 2);
-        Assert.assertTrue(serverAbilityControlManager.getServerNotSupportAbility().contains("test-001231231"));
-    
-        // remove then support
-        serverAbilityControlManager.removeTable("test-001231");
-        Assert.assertEquals(serverAbilityControlManager.isClusterEnableAbilityNow(AbilityKey.TEST_1), AbilityStatus.UNKNOWN);
-        serverAbilityControlManager.removeTable("test-001231231");
-        Assert.assertEquals(serverAbilityControlManager.isClusterEnableAbilityNow(AbilityKey.TEST_1), AbilityStatus.NOT_SUPPORTED);
-    }
-    
-    @Test
-    public void testClientRemove() {
-        Map<AbilityKey, Boolean> clientTa = new HashMap<>();
-        clientTa.put(AbilityKey.TEST_2, true);
-        clientTa.put(AbilityKey.TEST_1, false);
-        AbilityTable clientTable = new AbilityTable();
-        clientTable.setConnectionId("test-01111");
-        clientTable.setAbility(clientTa);
-        clientTable.setServer(true);
-        serverAbilityControlManager.addNewTable(clientTable);
-        Assert.assertTrue(serverAbilityControlManager.contains(clientTable.getConnectionId()));
-        serverAbilityControlManager.removeTable("test-01111");
-        Assert.assertFalse(serverAbilityControlManager.contains(clientTable.getConnectionId()));
-    }
-
     @Test
     public void testComponent() throws InterruptedException {
         enabled = 0;
@@ -177,7 +85,7 @@ public class AbilityControlManagerTest {
     
     @Test
     public void testCurrentNodeAbility() {
-        Set<AbilityKey> keySet = serverAbilityControlManager.getCurrentRunningAbility().keySet();
+        Set<AbilityKey> keySet = serverAbilityControlManager.getCurrentNodeAbilities().keySet();
         // diable all
         keySet.forEach(key -> serverAbilityControlManager.disableCurrentNodeAbility(key));
         // get all
@@ -190,42 +98,6 @@ public class AbilityControlManagerTest {
         keySet.forEach(key -> {
             Assert.assertTrue(serverAbilityControlManager.isCurrentNodeAbilityRunning(key));
         });
-        
-        // add node doesn't support ability table
-        AbilityTable abilityTable = new AbilityTable();
-        abilityTable.setServer(true);
-        abilityTable.setConnectionId("adsadsa1");
-        serverAbilityControlManager.addNewTable(abilityTable);
-        // cluster abilities close
-        Assert.assertEquals(serverAbilityControlManager.isClusterEnableAbilityNow(AbilityKey.TEST_1), AbilityStatus.UNKNOWN);
-        
-        AbilityTable abilityTable1 = new AbilityTable();
-        abilityTable1.setServer(true);
-        abilityTable1.setConnectionId("adsadsa2");
-        serverAbilityControlManager.addNewTable(abilityTable1);
-        // cluster abilities still close
-        Assert.assertEquals(serverAbilityControlManager.isClusterEnableAbilityNow(AbilityKey.TEST_1), AbilityStatus.UNKNOWN);
-        Assert.assertNull(serverAbilityControlManager.getClusterAbility());
-    
-        AbilityTable abilityTable2 = new AbilityTable();
-        abilityTable2.setServer(true);
-        abilityTable2.setConnectionId("adsadsa3");
-        Map<AbilityKey, Boolean> clientTa = new HashMap<>();
-        clientTa.put(AbilityKey.TEST_2, true);
-        clientTa.put(AbilityKey.TEST_1, false);
-        abilityTable2.setAbility(clientTa);
-        serverAbilityControlManager.addNewTable(abilityTable2);
-        // cluster abilities still close
-        Assert.assertEquals(serverAbilityControlManager.isClusterEnableAbilityNow(AbilityKey.TEST_1), AbilityStatus.UNKNOWN);
-        Assert.assertNull(serverAbilityControlManager.getClusterAbility());
-        
-        // remove
-        serverAbilityControlManager.removeTable("adsadsa3");
-        serverAbilityControlManager.removeTable("adsadsa2");
-        serverAbilityControlManager.removeTable("adsadsa1");
-        // cluster abilities open
-        Assert.assertEquals(serverAbilityControlManager.isClusterEnableAbilityNow(AbilityKey.TEST_1), AbilityStatus.SUPPORTED);
-        Assert.assertNotNull(serverAbilityControlManager.getClusterAbility());
     }
     
     class TestHandlerMapping implements HandlerMapping {
