+++ core/src/test/java/com/alibaba/nacos/core/ability/config/TestAbilityConfig.java
@@ -17,7 +17,6 @@
 package com.alibaba.nacos.core.ability.config;
 
 import com.alibaba.nacos.api.ability.constant.AbilityKey;
-import com.alibaba.nacos.common.ability.inter.AbilityHandlerRegistry;
 
 import java.util.Set;
 
@@ -32,9 +31,5 @@ public class TestAbilityConfig extends AbilityConfigs {
         Set<AbilityKey> serverAbilityKeys = super.getServerAbilityKeys();
         serverAbilityKeys.add(AbilityKey.TEST_1);
         serverAbilityKeys.add(AbilityKey.TEST_2);
-    }
-    
-    public void setAbilityControlManager(AbilityHandlerRegistry abilityHandlerRegistry) {
-        super.setAbilityHandlerRegistry(abilityHandlerRegistry);
     }
 }
